refacto: move publish logic in a dedicated hook
