+++ contexts/bal-data.js
@@ -37,7 +37,7 @@ export const BalDataContextProvider = React.memo(({
 
   const {token} = useContext(TokenContext)
 
-  const [habilitation, reloadHabilitation, isHabilitationValid, habilitationIsLoading] = useHabilitation(initialBaseLocale, token)
+  const [habilitation, reloadHabilitation, isHabilitationValid, habilitationIsLoading, isHabilitationDisplayed, setIsHabilitationDisplayed] = useHabilitation(initialBaseLocale, token)
 
   const reloadParcelles = useCallback(async () => {
     const parcelles = await getParcelles(baseLocale._id)
@@ -177,7 +177,9 @@ export const BalDataContextProvider = React.memo(({
     setVoie,
     setToponyme,
     certifyAllNumeros,
-    habilitationIsLoading
+    habilitationIsLoading,
+    isHabilitationDisplayed,
+    setIsHabilitationDisplayed
   }), [
     isEditing,
     editingId,
@@ -208,7 +210,9 @@ export const BalDataContextProvider = React.memo(({
     certifyAllNumeros,
     isRefrehSyncStat,
     refreshBALSync,
-    habilitationIsLoading
+    habilitationIsLoading,
+    isHabilitationDisplayed,
+    setIsHabilitationDisplayed
   ])
 
   return (
